pyproject.toml
[tool.poetry]
name = "backend-with"
version = "0.1.0"
description = ""
authors = ["dapt <diegodapt18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytelegrambotapi = "^4.13.0"
django = "^4.2.5"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = 'python manage.py runserver'
ngrok = 'ngrok http 8000'
